+++ CHANGELOG.md
@@ -6,7 +6,7 @@ List all changes in various categories:
 * Bug fixes
 * Known issues
 
-## 3.0.0 (2018-11-03) ##
+## 3.0.0 (2018-11-13) ##
 
 ### Release notes ###
 
@@ -15,7 +15,7 @@ List all changes in various categories:
 
 ### Bug fixes ###
 
-- Update requests dependency to latest version (2.20.0)
+- Update requests dependency to latest version (2.20.1)
 - [OAuth2] Remove unused redirect_uri_port_availability_timeout parameter
 
 ## 2.0.0 (2018-10-09) ##